add change roles on members

+++ app/Http/Controllers/ProjectController.php
@@ -148,6 +148,9 @@ class ProjectController extends Controller
 
     public function leaveProject($id) {
         $project_user = ProjectUser::find(['user_id' => Auth::id(), 'project_id' => $id]);
+        if ($project_user->user_role == "MANAGER" && Project::find($id)->managers()->count() < 2) {
+            return redirect()->back()->withErrors("You can't leave the project because you are the only manager of the project");
+        }
         $project_user->delete();
         return redirect('/projects');
     }
